Moved files to infra folder in user service
