+++ user-service/main.go
@@ -6,10 +6,10 @@ import (
 	"github.com/gin-contrib/cors"
 	"github.com/gin-gonic/gin"
 	"github.com/sirupsen/logrus"
-	"github.com/vadgun/gotrelloclone/user-service/config"
 	"github.com/vadgun/gotrelloclone/user-service/handlers"
-	"github.com/vadgun/gotrelloclone/user-service/logger"
-	"github.com/vadgun/gotrelloclone/user-service/metrics"
+	"github.com/vadgun/gotrelloclone/user-service/infra/config"
+	"github.com/vadgun/gotrelloclone/user-service/infra/logger"
+	"github.com/vadgun/gotrelloclone/user-service/infra/metrics"
 	"github.com/vadgun/gotrelloclone/user-service/repositories"
 	"github.com/vadgun/gotrelloclone/user-service/routes"
 	"github.com/vadgun/gotrelloclone/user-service/services"
